+++ src/canister/src/storage/history_point_storage.rs
@@ -1,3 +1,4 @@
+use canister_types::models::api_error::ApiError;
 use ic_stable_structures::memory_manager::MemoryId;
 
 use super::{
@@ -16,5 +17,14 @@ impl CellStorage<u64> for HistoryPointStorage {
 
     fn memory_id() -> MemoryId {
         HISTORY_POINT_MEMORY_ID
+    }
+}
+
+impl HistoryPointStorage {
+    pub fn get_next() -> Result<u64, ApiError> {
+        let current = Self::get().unwrap_or(1);
+        let next = current + 1;
+        Self::set(next)?;
+        Ok(next)
     }
 }